gitleaks: allow skipping patterns download in offline
In offline mode if the patterns file is considered outdated, a
warning is printed. If the patterns file doesn't exist then the
tool cannot run. It is assumed that the user has setup some
background job to synchronize new versions of the patterns file,
to enable the commit checking to run entirely offline.

A grace period of 1 week is permitted for outdated patterns
file, but over this limit the patterns are considered stale
and this is treated as a blocking error.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>