gitleaks: run offline if LEAKTK_SCANNER_AUTOFETCH=0 env is set
The new LEAKTK_SCANNER_AUTOFETCH=0 environment variable can not
be used to control whether rh-gitleaks will try to download
the gitleaks binary or refresh the pattern file, when scanning
a git commit.

An env variable is used rather than a CLI arg, because this
allows the developer to override it easily.

For example

 $ LEAKTK_SCANNER_AUTOFETCH=0 git commit

Or this can be added to a .bashrc file to make it permanent,
with 'LEAKTK_SCANNER_AUTOFETCH=1' then being usable to
selectively turn on refresh.

Fixes: https://gitlab.corp.redhat.com/infosec-public/developer-workbench/tools/-/issues/14
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>